Fix implicit narrowing conversion flagged by CodeQL
Changed totalScenarios from int to long to match Stream.count()
return type, avoiding implicit long-to-int narrowing.

+++ src/main/java/com/heymumford/ftoc/formatter/TocFormatter.java
@@ -532,7 +532,7 @@ public class TocFormatter {
         html.append(buildHtmlFilterMetadata(
             includeTags, excludeTags));
 
-        int totalScenarios = 0;
+        long totalScenarios = 0;
         for (Feature feature : features) {
             totalScenarios += feature
                 .getFilteredScenarios(includeTags, excludeTags)
